moved plan and tf scan functionality into base command
set plan scanning to be triggered by --plan variable flag
